+++ kinto/plugins/admin/__init__.py
@@ -1,5 +1,6 @@
 from pyramid.static import static_view
 from pyramid.httpexceptions import HTTPTemporaryRedirect
+from .views import admin_home_view
 
 
 def includeme(config):
@@ -8,10 +9,13 @@ def includeme(config):
     # Expose capability.
     config.add_api_capability(
         "admin",
-        version="1.8.0",
+        version="1.8.1",
         description="Serves the admin console.",
         url="https://github.com/Kinto/kinto-admin/",
     )
+
+    config.add_route('admin_home', '/admin/')
+    config.add_view(admin_home_view, route_name="admin_home")
 
     build_dir = static_view('kinto.plugins.admin:build', use_subpath=True)
     config.add_route('catchall_static', '/admin/*subpath')